Custom hook for SearchScreen
